Fix the issue when calculating score from classification model (#9)
Bump the version to 0.1.1

Now the customized sklearn estimator is only compatible to
calculate the score by a regression PyTorch model. In this PR,
another argument `target` is introduced to the class object, which
can indicate if the model is a classifier or a regressor, and
accordingly the corresponding score can be calculated.