Fix the issue when hyperparameter tuning by ray (#29)
In the current version, when tuning hyperparameters the
earlystopping instance will be as one of inputs for tuning
function, which will be reused among different trials, then this
made many trials stop unexpectedly, since the earlystopping
instance is not initialized anymore after the first trial.

In this PR, this issue is fixed by defining the initialized
earlystopping instance inside tuning function. Then for each trial
there will be an independent earlystopping instance.