Convert the classification model to genenric model
Since the difference between classification and regression models
are only output data, where classifier outputs probability and
regressor outputs regressive values. Thus, the model structures
for the both purposes can be identical except the final output
layer, which can be controlled easily by the parameters.

Then the codebase can be simplified a lot and there are also
generic models for classification and regression.